feat: port to windows and macos
For both OSes the rados, cephfs, and rdb benchmarks are not included.
The point of the port is to have the file target, which can run on a
manually mointed SMB share.

The new FileDescriptor type abstracts the distinction between UNIX
file descriptors and Windows file handles.

Build tags are used to select OS implementations.